Desktop/Models-Career-API/Models-Hossam-API/requirements.txt
streamlit
together
pdfplumber
python-dotenv
Flask 
resume-parser 
